feat: refactor hygraph queries
