+++ src/app/lib/hygraph/pages.js
@@ -1,0 +1,38 @@
+import { gql } from '@apollo/client';
+import client from '@/app/lib/apollo-client';
+
+export const getPageBySlug = async (slug) => {
+    const { data, errors } = await client.query({
+        query: gql`
+            query GetPageBySlug($slug: String!) {
+                page(where: { slug: $slug }) {
+                  slug
+                  id
+                  title
+                  content {
+                    raw
+                    html
+                    markdown
+                    text
+                  }
+                   seoOverride {
+                    description
+                    title
+                    image {
+                      altText
+                      url
+                    }
+                  }   
+                }
+              }
+        `,
+        variables: { slug },
+    });
+
+    if (errors) {
+        const error = apolloError;
+        throw new Error("Failed to fetch post.");
+    }
+
+    return data.page
+}
