+++ src/pages/about/team/index.js
@@ -2,7 +2,8 @@ import SEOHead from '@/app/components/SEOHead';
 import HeroSection from '@/app/components/about/team/HeroSection';
 import JobOpeningsSection from '@/app/components/about/team/JobOpeningsSection';
 import TeamGridSection from '@/app/components/about/team/TeamGridSection';
-import { getAllEmployees, getPageBySlug } from '@/app/lib/hygraph';
+import { getAllEmployees } from '@/app/lib/hygraph/employees';
+import { getPageBySlug } from '@/app/lib/hygraph/pages';
 
 function TeamPage({ page, employees }) {
